requirements.txt
Flask==3.1.0
openai==1.63.0
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
regex